Move files to the correct location
Also adds a maven check so 2.2 does not build if 2.1 files are left around
