+++ community/cypher/pom.xml
@@ -144,6 +144,31 @@
           </excludes>
         </configuration>
       </plugin>
+
+      <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-enforcer-plugin</artifactId>
+          <version>1.3.1</version>
+          <executions>
+              <execution>
+                  <id>enforce-dont-exist</id>
+                  <goals>
+                      <goal>enforce</goal>
+                  </goals>
+                  <configuration>
+                      <rules>
+                          <requireFilesDontExist>
+                              <files>
+                                  <file>cypher-compiler-2.1</file>
+                              </files>
+                          </requireFilesDontExist>
+                      </rules>
+                      <fail>true</fail>
+                  </configuration>
+              </execution>
+          </executions>
+      </plugin>
+
     </plugins>
 
     <pluginManagement>
